+++ src/main/java/com/bew/demo/service/DictamenService.java
@@ -12,7 +12,7 @@ public interface DictamenService {
 	void saveDictamen(DictamenDTO dictamenDTO);
 	void updateDictamen(DictamenDTO dictamenDTO)throws EmptyResultException;
 	void deleteDictamen(Integer idDictamen)throws EmptyResultException;
-	DictamenDTO findByIdAlumno(Integer idAlumno);
+	DictamenDTO findByIdAlumno(Long idAlumno);
 	List<DictamenDTO> findByEstado(String estado);
 	
 }
